:art: デザイン関係の定数のファイル構成を見直し
