+++ src/scenes/MainMenu/MainMenuScene.ts
@@ -1,6 +1,8 @@
 import * as Phaser from 'phaser';
-import { SCREEN, PADDING } from 'src/consts/layout';
-import { COLORS, LARGE_TEXT_STYLE, DEFAULT_TEXT_STYLE } from 'src/consts/styles';
+import { LAYOUT } from '@/consts/styles/layout';
+import { PADDING } from '@/consts/styles/components';
+import { COLORS } from '@/consts/styles/color';
+import { LARGE_TEXT_STYLE, DEFAULT_TEXT_STYLE } from '@/consts/styles/text';
 import { SCENE_KEYS } from 'src/consts/scenes';
 import i18next from 'src/i18n/config';
 
@@ -14,14 +16,14 @@ export class MainMenuScene extends Phaser.Scene {
     this.cameras.main.setBackgroundColor(COLORS.BACKGROUND);
 
     // タイトルテキスト
-    this.add.text(SCREEN.CENTER_X, 200, i18next.t('title'), {
+    this.add.text(LAYOUT.SCREEN.CENTER_X, 200, i18next.t('title'), {
       ...LARGE_TEXT_STYLE,
       color: COLORS.TEXT.PRIMARY
     }).setOrigin(0.5);
 
     // スタートボタン
     const startButton = this.add.text(
-      SCREEN.CENTER_X,
+      LAYOUT.SCREEN.CENTER_X,
       300,
       i18next.t('startButton'),
       {
@@ -36,7 +38,7 @@ export class MainMenuScene extends Phaser.Scene {
     // 言語切り替えボタン
     // TODO: ブラウザ等の設定から自動で反映されるようにしたい。
     this.add.text(
-      SCREEN.CENTER_X + 200, // 画面中央から右に200px
+      LAYOUT.SCREEN.CENTER_X + 200, // 画面中央から右に200px
       50,
       i18next.language === 'ja' ? '🇯🇵' : '🇺🇸',
       DEFAULT_TEXT_STYLE
@@ -55,7 +57,7 @@ export class MainMenuScene extends Phaser.Scene {
     // ボタンのクリックイベント
     startButton.on('pointerdown', () => {
       // TODO: プレイヤー選択画面を廃止しする
-      
+
       // プレイヤー選択画面を廃止し、マッチメイキング画面に直接遷移
       this.scene.start(SCENE_KEYS.MATCHMAKING);
     });
